Use explicit casting for consistent float calculations.
Replaced implicit conversions with explicit `static_cast<float>` to ensure consistent and clear float calculations. Adjusted parameter types for `draw_texture_to_fill_rect` to better align with expected inputs.

+++ src/ui/components/component_base.cpp
@@ -2,22 +2,22 @@
 
 #include <algorithm>
 
-void artifact::ComponentBase::draw_texture_to_fill_rect(const float width, const float height, const float x, const float y, const Texture2D &texture)
+void artifact::ComponentBase::draw_texture_to_fill_rect(const int width, const int height, const int x, const int y, const Texture2D &texture)
 {
     // Calculate scaling factors for both dimensions
-    const float scaleX = width / texture.width;
-    const float scaleY = height / texture.height;
+    const float scaleX = static_cast<float>(width) / static_cast<float>(texture.width);
+    const float scaleY = static_cast<float>(height) / static_cast<float>(texture.height);
 
     // Use the larger scale to ensure the image covers the whole screen
     const float scale = std::max(scaleX, scaleY);
 
     // Calculate the scaled dimensions
-    const float scaledWidth = texture.width * scale;
-    const float scaledHeight = texture.height * scale;
+    const float scaledWidth = static_cast<float>(texture.width) * scale;
+    const float scaledHeight = static_cast<float>(texture.height) * scale;
 
     // Calculate position to center the image
-    const float dx = (width - scaledWidth) * 0.5f;
-    const float dy = (height - scaledHeight) * 0.5f;
+    const float dx = (static_cast<float>(width) - scaledWidth) * 0.5f;
+    const float dy = (static_cast<float>(height) - scaledHeight) * 0.5f;
 
     // Define source rectangle (the entire texture)
     const Rectangle source = {0.0f, 0.0f, static_cast<float>(texture.width), static_cast<float>(texture.height)};
@@ -26,5 +26,5 @@ void artifact::ComponentBase::draw_texture_to_fill_rect(const float width, const
     const Rectangle dest = {dx, dy, scaledWidth, scaledHeight};
 
     // Draw the texture
-    DrawTexturePro(texture, source, dest, {x, y}, 0.0f, WHITE);
+    DrawTexturePro(texture, source, dest, {static_cast<float>(x), static_cast<float>(y)}, 0.0f, WHITE);
 }
